Struct snowcap_bencher::BencherArguments

pub struct BencherArguments {
    pub bench_type: BencherType,
    pub iterations: usize,
    pub max_time: u64,
    pub ignore_nan: bool,
    pub random: bool,
    pub tree: bool,
    pub main: bool,
    pub mif: bool,
    pub mil: bool,
    pub global_optimum: bool,
    pub threads: Option<usize>,
    pub output_csv: Option<String>,
    pub output_json: Option<String>,
}

Arguments required for the bencher

Fields
bench_type: BencherType

Type of benchmark

iterations: usize

Number of iterations to repeat

max_time: u64

Maximum allowed time per run, in seconds. If the time exceeds the limit, then the cost will be set to nan.

ignore_nan: bool

If this flag is set, then nan values will be ignored, and not added to the resulting csv file

random: bool

Perform benching the random baseline

tree: bool

Perform benching the tree strategy

main: bool

Perform benching the main strategy

mif: bool

Perform benching the most-important-first baseline strategy

mil: bool

Perform benching the most-important-first baseline strategy

global_optimum: bool

search for the global optimum

threads: Option<usize>

Number of threads to use. Defaults to the number of threads available on the system.

output_csv: Option<String>

Output file to store the results. Two different files will be created: "NAME_strategy.csv", "NAME_tree.csv" and "NAME_random.csv"! Don't provide the file ending ".csv"!

output_json: Option<String>

Output file to store the results in json format. Give the entire path, including the json ending.